refactor: separate Office worker contracts
